+++ src/degree_to_radian.F90
@@ -3,6 +3,7 @@
 !*************************************************************
 module DEGREE_TO_RADIAN
 use NAMELIST
+use, intrinsic :: iso_fortran_env
 implicit none
 public :: deg_to_rad
 private
@@ -12,8 +13,8 @@ contains
 ! Function convert angle in degrees to radians
 !*********************************************
 function deg_to_rad(angle) result(ans)
-    real, intent(in) :: angle
-    real :: ans
+    real(real64), intent(in) :: angle
+    real(real64) :: ans
     ans = angle*pi/180
 end function deg_to_rad
