+++ src/height_of_slab.F90
@@ -3,13 +3,14 @@
 !********************************************
 module HEIGHT_OF_SLAB
 use NAMELIST
+use, intrinsic :: iso_fortran_env
 implicit none
 public:: h_slab
 private
 contains
 function h_slab(h) result(ans)
-    integer :: h
-    real :: ans
+    integer(int64) :: h
+    real(real64) :: ans
     if (h == 1) then !using index 1 for surface
         ans = H_S
     elseif (h == 2) then ! using index 2 for deep
